users refactoring
