+++ backend_api/app/applications/auth/auth_handler.py
@@ -14,7 +14,9 @@ class AuthHandler:
         self.algorithm = settings.JWT_ALGORITHM
 
     async def get_login_token_pairs(self, data: OAuth2PasswordRequestForm, session: AsyncSession):
-        pass
+        user_email = data.username
+        user_password = data.password
+
 
 
 auth_handler = AuthHandler()
